pyproject.toml
[build-system]
requires = ["setuptools>=65.5.1", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "invoke-agent"
version         = "0.1.0"
description     = "A lightweight API execution layer for LLM agents."
readme          = "README.md"
requires-python = ">=3.8"
license         = { text = "MIT" }
authors         = [ { name = "mercury0100" } ]

# Inline tables must use bare keys, not quoted ones:
urls = { Homepage = "https://github.com/mercury0100/invoke", Issues = "https://github.com/mercury0100/invoke/issues" }

classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

dependencies = [
  "requests",
  "python-dotenv",
  "tldextract",
  "pyyaml",
  "langchain-core",
  "langchain"
]

[tool.setuptools.packages.find]
where   = ["."]
include = ["invoke_agent*"]
